+++ src/renderer/App.tsx
@@ -13,10 +13,12 @@ const APIMsg = {
 const styles : {[key: string]:React.CSSProperties} = {
   dropAreaNormal:{
     //background:"#000"
+    transition: "0.1s",
   },
   dropAreaOver:{
     //background:"#f00",
     //width:"80%",
+    transition: "0.3s",
     transform:"scale(1.1,1.1)"
   }
 }
@@ -30,6 +32,7 @@ const Home:React.FC = () => {
 
   useKey(['ArrowLeft'],()=>{imagePrevious()});
   useKey(['ArrowRight'],()=>{imageNext()});
+  useKey(['Delete'],()=>{imageDelete()});
 
   const openImage1Dir = () =>{
     myAPI.openDir(APIMsg.IMAGES1);
@@ -38,19 +41,30 @@ const Home:React.FC = () => {
     myAPI.openDir(APIMsg.IMAGES2);
   }
 
-  const imageNext = () =>{
-    (imgNum>0)?setImgNum(imgNum-1):setImgNum(images1.length-1);
+  const imagePrevious  = () =>{
+    if(images1.length > images2.length){
+      (imgNum>0)?setImgNum(imgNum-1):setImgNum(images1.length-1);
+    }else{
+      (imgNum>0)?setImgNum(imgNum-1):setImgNum(images2.length-1);
+    }
   }
-  const imagePrevious =() =>{
-    (imgNum<images1.length-1)?setImgNum(imgNum+1):setImgNum(0);
+  const imageNext =() =>{
+    if(images1.length > images2.length){
+      (imgNum<images1.length-1)?setImgNum(imgNum+1):setImgNum(0);
+    }else{
+      (imgNum<images2.length-1)?setImgNum(imgNum+1):setImgNum(0);
+    }
   }
 
   const imageDelete =()=>{
-    myAPI.deleteImage(images1[imgNum],images2[imgNum]);
-    images1.splice(imgNum,1);
-    images2.splice(imgNum,1);
-    setImages1([...images1]);
-    setImages2([...images2]);
+    var result = window.confirm('削除してもよろしいでしょうか?');
+    if(result){
+      myAPI.deleteImage(images1[imgNum],images2[imgNum]);
+      images1.splice(imgNum,1);
+      images2.splice(imgNum,1);
+      setImages1([...images1]);
+      setImages2([...images2]);
+    }
   }
 
   const dropFolder1 = (e:React.DragEvent) =>{
@@ -88,16 +102,12 @@ const Home:React.FC = () => {
       return(
       <div className='flexdwrap'>
         <div id='editbar'>
-        <Segment>
-            <Button className='editbutton' icon="arrow alternate circle left outline" onClick={()=>{imageNext()}}></Button>
+        <Segment inverted>
+            <Button className='editbutton' icon="arrow alternate circle left outline" onClick={()=>{imagePrevious()}}/>
             <Button className='editbutton' icon="folder open outline" onClick={()=>{openImage1Dir()}}/>
-            <Button className='editbutton' icon="trash alternate outline" onClick={()=>{
-              var result = window.confirm('削除してもよろしいでしょうか?');
-              if(result)imageDelete();
-              }}>  
-            </Button>
+            <Button className='editbutton' icon="trash alternate outline" onClick={()=>{imageDelete();}}/>  
             <Button className='editbutton' icon="folder open outline" onClick={()=>{openImage2Dir()}}/>
-            <Button className='editbutton' icon="arrow alternate circle right outline" onClick={()=>{imagePrevious()}}></Button>
+            <Button className='editbutton' icon="arrow alternate circle right outline" onClick={()=>{imageNext()}}/>
           </Segment>
         </div>
       </div>);
